Add secret filters and improve UI
You can now see your secret or only the shared ones.
Account creation need password confirmation.
Add secret and Add folder are more visible.

+++ src/components/secrets/SecretListItem/Options.js
@@ -14,42 +14,56 @@ const propTypes = {
 
 function SecretListItemOptions({ parentFolderId, secret }) {
   const currentUser = AppUIStore.getCurrentUser();
+  const canShare = secret.canBeSharedBy(currentUser);
+  if (!canShare && secret.type === 'folder') {
+    return null;
+  }
   return (
     <Dropdown id="secret-action" pullRight>
       <Dropdown.Toggle>
         <Icon id="more-vert" size="small" />
       </Dropdown.Toggle>
       <Dropdown.Menu>
-        <Dropdown.MenuItem
-          onSelect={() => ShowSecretUIActions.showSecret({ secret, tab: secret.type === 'folder' ? 'access' : 'details' })}
-        >
-          Show
-        </Dropdown.MenuItem>
-        <Dropdown.MenuItem
-          onSelect={() => ShowSecretUIActions.showSecret({ secret, tab: 'access' })}
-          disabled={!secret.canBeSharedBy(currentUser)}
-        >
-          Share
-        </Dropdown.MenuItem>
+        {
+          secret.type !== 'folder' &&
+            <Dropdown.MenuItem
+              onSelect={() => ShowSecretUIActions.showSecret({ secret, tab: secret.type === 'folder' ? 'access' : 'details' })}
+            >
+              Show
+            </Dropdown.MenuItem>
+        }
+        {
+          canShare &&
+          (
+            <div>
+              <Dropdown.MenuItem
+                onSelect={() => ShowSecretUIActions.showSecret({ secret, tab: 'access' })}
+              >
+                Share
+              </Dropdown.MenuItem>
 
-        <Dropdown.MenuItem divider />
+              <Dropdown.MenuItem divider />
 
-        <Dropdown.MenuItem
-          onSelect={() => MetadataActions.removeSecretFromCurrentFolder({
-            secret,
-            currentFolderId: parentFolderId,
-          })}
-          disabled={!parentFolderId || !secret.canBeUpdatedBy(currentUser)}
-        >
-          Remove from this folder
-        </Dropdown.MenuItem>
+              {
+                parentFolderId &&
+                  <Dropdown.MenuItem
+                    onSelect={() => MetadataActions.removeSecretFromCurrentFolder({
+                      secret,
+                      currentFolderId: parentFolderId,
+                    })}
+                  >
+                    Remove from this folder
+                  </Dropdown.MenuItem>
+              }
 
-        <Dropdown.MenuItem
-          onSelect={() => MetadataActions.deleteSecret({ secret })}
-          disabled={!secret.canBeUpdatedBy(currentUser)}
-        >
-          Delete
-        </Dropdown.MenuItem>
+              <Dropdown.MenuItem
+                onSelect={() => MetadataActions.deleteSecret({ secret })}
+              >
+                Delete
+              </Dropdown.MenuItem>
+            </div>
+          )
+        }
       </Dropdown.Menu>
     </Dropdown>
   );
